feat: Don't add redundant elements to output
Code has been changed to find direct and indirect dependencies of a definition element.
Dependencies of a definition element are added to the output only if the element itself is to be added to the output.
This will ensure that in the output file all elements are referenced by a definition element either directly or indirectly.
Associated test cases have been added, and/or updated.