feat: Improve OVAL ID replacement logic to replace only duplicate IDs
This commit optimizes the logic to replace OVAL IDs from input
files when generating the output XML. Currently, all OVAL IDs are
regenerated with an incrementing suffix to mitigate duplicate IDs across
input files.
This has the overhead of generating a large number of IDs and replacing
all of them one by one in the output XML.
This logic has been optimized to regenerate and replace new IDs in the
output XML only for duplicate IDs.